pyproject.toml
[tool.poetry]
name = "geckopackage"
version = "0.1.0"
description = "Gecko dependencies"
authors = ["Ansor T. <ansor.tagaev101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
httpx = {extras = ["cli"], version = "^0.28.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
